feat: Refactor date and time sheet view models
- Create separate view models for date and time sheet interactions
- Implement DateSheetViewModel and TimeSheetViewModel with shared functionality
- Update task sheet views to use new view models
- Remove redundant code and improve state management for date/time selection
